rename ZipscriptVFSDataSFV package to match path
