+++ src/plugins/zipscript/zipscript-master/src/main/java/org/drftpd/zipscript/master/sfv/hooks/ZipscriptPostHook.java
@@ -47,8 +47,8 @@ import org.drftpd.master.vfs.FileHandle;
 import org.drftpd.zipscript.common.sfv.SFVInfo;
 import org.drftpd.zipscript.common.sfv.SFVStatus;
 import org.drftpd.zipscript.master.sfv.SFVTools;
-import org.drftpd.zipscript.master.sfv.ZipscriptVFSDataSFV;
 import org.drftpd.zipscript.master.sfv.event.SFVMemberTransferEvent;
+import org.drftpd.zipscript.master.sfv.vfs.ZipscriptVFSDataSFV;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
